+++ src/main/java/tk/cth451/transitrailmod/blocks/HungArrowSign.java
@@ -1,6 +1,5 @@
 package tk.cth451.transitrailmod.blocks;
 
-import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.properties.IProperty;
 import net.minecraft.block.properties.PropertyDirection;
@@ -17,53 +16,19 @@ import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 import tk.cth451.transitrailmod.TransitRailMod;
+import tk.cth451.transitrailmod.blocks.prototype.ArrowSign;
 import tk.cth451.transitrailmod.enums.EnumArrow;
 import tk.cth451.transitrailmod.init.ModItems;
 
-public class HungArrowSign extends Block {
-	
-	public static final PropertyDirection FACING = PropertyDirection.create("facing", EnumFacing.Plane.HORIZONTAL);
-	public static final PropertyEnum ARROW = PropertyEnum.create("arrow", EnumArrow.class);
-	
+public class HungArrowSign extends ArrowSign {
+		
 	public HungArrowSign(Material materialIn) {
 		super(Material.iron);
-		this.setLightLevel(1.0F);
 		this.setUnlocalizedName("hung_arrow_sign");
 		this.setCreativeTab(TransitRailMod.tabTransitRail);
-		this.setDefaultState(getDefaultState()
-				.withProperty(ARROW, EnumArrow.ARROW_UP)
-				.withProperty(FACING, EnumFacing.NORTH));
-	}
-	
-	// Block state
-	@Override
-	protected BlockState createBlockState() {
-		return new BlockState(this, new IProperty[] {ARROW, FACING});
-	}
-	
-	// meta 1122
-	// 11: ARROW
-	// 22: FACING
-	@Override
-	public IBlockState getStateFromMeta(int meta) {
-		EnumFacing pFacing = EnumFacing.getHorizontal(meta % 4);
-		EnumArrow pArrow = EnumArrow.fromMeta(meta / 4);
-		return this.getDefaultState().withProperty(ARROW, pArrow).withProperty(FACING, pFacing);
-	}
-	
-	@Override
-	public int getMetaFromState(IBlockState state) {
-		int mFacing = ((EnumFacing) state.getValue(FACING)).getHorizontalIndex();
-		int mArrow = ((EnumArrow) state.getValue(ARROW)).toMeta();
-		return mArrow * 4 + mFacing;
 	}
 	
 	// Properties
-	@Override
-	public boolean isTranslucent() {
-		return true;
-	}
-	
 	@Override
 	public boolean isPassable(IBlockAccess worldIn, BlockPos pos) {
 		return true;
@@ -76,22 +41,10 @@ public class HungArrowSign extends Block {
     }
 	
 	@Override
-	public boolean isOpaqueCube() {
-        return false;
-    }
-	
-	@Override
 	public boolean isFullCube()
     {
         return false;
     }
-	
-	@Override
-	@SideOnly(Side.CLIENT)
-	public EnumWorldBlockLayer getBlockLayer()
-	{
-		return EnumWorldBlockLayer.TRANSLUCENT;
-	}
 	
 	// Appearance
 	@Override
@@ -102,25 +55,5 @@ public class HungArrowSign extends Block {
 		} else {
 			this.setBlockBounds(0.4375F, 0.5F, 0.0F, 0.5625F, 1.0F, 1.0F);
 		}
-	}
-	
-	// Interactions
-	@Override
-	public IBlockState onBlockPlaced(World worldIn, BlockPos pos, EnumFacing facing, float hitX, float hitY, float hitZ, int meta, EntityLivingBase placer){
-		// set facing to the direction player is facing
-		IBlockState state = super.onBlockPlaced(worldIn, pos, facing, hitX, hitY, hitZ, meta, placer);
-		EnumFacing thisFacing = placer.getHorizontalFacing();
-		return this.getActualState(state, worldIn, pos).withProperty(FACING, thisFacing);
-	}
-	
-	@Override
-	public boolean onBlockActivated(World worldIn, BlockPos pos, IBlockState state, EntityPlayer playerIn,
-			EnumFacing side, float hitX, float hitY, float hitZ) {
-		if (playerIn.getHeldItem() != null) {
-			if (playerIn.getHeldItem().getItem() == ModItems.style_changer){
-				worldIn.setBlockState(pos, state.cycleProperty(ARROW));
-			}
-		}
-		return true;
 	}
 }