+++ src/main/java/tk/cth451/transitrailmod/init/ModBlocks.java
@@ -12,6 +12,7 @@ import tk.cth451.transitrailmod.blocks.ClosedPlatformPanelBlock;
 import tk.cth451.transitrailmod.blocks.ClosedPlatformTop;
 import tk.cth451.transitrailmod.blocks.HungArrowSign;
 import tk.cth451.transitrailmod.blocks.LogoBlock;
+import tk.cth451.transitrailmod.blocks.PlatformArrowSign;
 import tk.cth451.transitrailmod.blocks.PlatformGateBlock;
 import tk.cth451.transitrailmod.blocks.PlatformPanelBlock;
 
@@ -23,6 +24,7 @@ public class ModBlocks {
 	public static Block platform_panel_block;
 	public static Block logo_block;
 	public static Block hung_arrow_sign;
+	public static Block platform_arrow_sign;
 	public static Block platform_sign;
 	
 	public static void init() {
@@ -33,6 +35,7 @@ public class ModBlocks {
 		platform_panel_block = new PlatformPanelBlock(Material.iron);
 		logo_block = new LogoBlock(Material.iron);
 		hung_arrow_sign = new HungArrowSign(Material.iron);
+		platform_arrow_sign = new PlatformArrowSign(Material.iron);
 		//platform_sign = new PlatformSign(Material.iron);
 	}
 	
@@ -44,6 +47,7 @@ public class ModBlocks {
 		GameRegistry.registerBlock(platform_panel_block, platform_panel_block.getUnlocalizedName().substring(5));
 		GameRegistry.registerBlock(logo_block, logo_block.getUnlocalizedName().substring(5));
 		GameRegistry.registerBlock(hung_arrow_sign, hung_arrow_sign.getUnlocalizedName().substring(5));
+		GameRegistry.registerBlock(platform_arrow_sign, platform_arrow_sign.getUnlocalizedName().substring(5));
 		//GameRegistry.registerBlock(platform_sign, platform_sign.getUnlocalizedName().substring(5));
 	}
 	
@@ -56,6 +60,8 @@ public class ModBlocks {
 		registerRender(logo_block);
 		registerRender(hung_arrow_sign);
 		registerInventoryRender(hung_arrow_sign);
+		registerRender(platform_arrow_sign);
+		registerInventoryRender(platform_arrow_sign);
 		//registerRender(platform_sign);
 	}